pyproject.toml
[tool.poetry]
name = "gstreamer-python-sample"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycairo = "^1.19.1"
PyGObject = "^3.36.1"
numpy = "^1.19.1"
gstreamer-python = {git = "https://github.com/jackersson/gstreamer-python.git", rev = "v0.0.2"}

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.5.2"
mypy = "^0.782"
flake8 = "^3.8.3"
flake8-black = "^0.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
